refactor: broke components into directories
